+++ jackson/src/main/java/org/linguafranca/pwdb/kdbx/jackson/JacksonSerializableDatabase.java
@@ -17,6 +17,7 @@ package org.linguafranca.pwdb.kdbx.jackson;
 
 import com.ctc.wstx.api.WstxInputProperties;
 import com.ctc.wstx.api.WstxOutputProperties;
+import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.databind.module.SimpleModule;
@@ -40,6 +41,12 @@ import java.util.Objects;
 
 public class JacksonSerializableDatabase implements SerializableDatabase {
 
+    /**
+     * By default, deserialization will fail if an unknown property is found. Historically,
+     * this has been because the file mapping is incomplete, rather than the incoming file being wrong.
+     * So use this feature with caution, if at all.
+      */
+    public static boolean FAIL_ON_UNKNOWN_PROPERTIES = true;
     public KeePassFile keePassFile;
     private StreamEncryptor encryptor;
 
@@ -64,6 +71,8 @@ public class JacksonSerializableDatabase implements SerializableDatabase {
     @Override
     public JacksonSerializableDatabase load(InputStream inputStream) throws IOException {
         XmlMapper mapper = new XmlMapper();
+        mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES,
+                JacksonSerializableDatabase.FAIL_ON_UNKNOWN_PROPERTIES);
         SimpleModule module = new SimpleModule();
         module.addDeserializer(PropertyValue.class, new ValueDeserializer(encryptor, propertyValueStrategy));
         mapper.registerModule(module);
@@ -137,6 +146,7 @@ public class JacksonSerializableDatabase implements SerializableDatabase {
         keePassFile.meta.binaries.add(newBin);
     }
 
+    // TODO this gets binary at index but does not get binary with ID
     @Override
     public byte[] getBinary(int index) {
         KeePassFile.Binary binary = keePassFile.meta.binaries.get(index);